Add unit tests and fix compiler warnings
- Add 35 unit tests (store, fuzzy, db, integration) using GLib test framework
- Add `make test` target with TAP output
- Fix GActionEntry missing padding initializer warning
- Fix g_unlink implicit declaration (add glib/gstdio.h include)
- Fix int-to-float implicit conversion in overlay click handler

All tests pass. Zero warnings from gcc and clang with strict flags.
cppcheck exhaustive + scan-build: zero findings.

+++ src/clipium-window.c
@@ -214,8 +214,8 @@ on_overlay_click(GtkGestureClick *gesture,
         return;
     }
 
-    int card_w = gtk_widget_get_width(GTK_WIDGET(self->card_box));
-    int card_h = gtk_widget_get_height(GTK_WIDGET(self->card_box));
+    float card_w = (float)gtk_widget_get_width(GTK_WIDGET(self->card_box));
+    float card_h = (float)gtk_widget_get_height(GTK_WIDGET(self->card_box));
 
     if (card_point.x < 0 || card_point.y < 0 ||
         card_point.x > card_w || card_point.y > card_h) {
